Rename to useRenderMarkdownAsHTML
