+++ packages/component/src/hooks/useRenderMarkdownAsHTML.js
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 import WebChatUIContext from '../WebChatUIContext';
 
-export default function useRenderMarkdown() {
+export default function useRenderMarkdownAsHTML() {
   return useContext(WebChatUIContext).renderMarkdown;
 }
